code/GitHubCopilotSideCar/code/gh-cli-blog-agent/sidecar/requirements.txt
fastapi
github-copilot-sdk
httpx
python-dotenv
uvicorn
